Additional refactoring of services and wcpm
