+++ query/security/wcpm/src/main/java/org/codice/ddf/admin/security/wcpm/actions/persist/SaveWhitelistContexts.java
@@ -14,9 +14,7 @@
 package org.codice.ddf.admin.security.wcpm.actions.persist;
 
 import static org.codice.ddf.admin.common.message.DefaultMessages.failedPersistError;
-import static org.codice.ddf.admin.security.wcpm.commons.ContextPolicyServiceProperties.POLICY_MANAGER_PID;
-import static org.codice.ddf.admin.security.wcpm.commons.ContextPolicyServiceProperties.getWhitelistContexts;
-import static org.codice.ddf.admin.security.wcpm.commons.ContextPolicyServiceProperties.whiteListToPolicyManagerProps;
+import static org.codice.ddf.admin.security.common.services.PolicyManagerServiceProperties.getWhitelistContexts;
 
 import java.util.List;
 
@@ -28,6 +26,7 @@ import org.codice.ddf.admin.common.fields.common.ContextPath;
 import org.codice.ddf.admin.configurator.Configurator;
 import org.codice.ddf.admin.configurator.ConfiguratorFactory;
 import org.codice.ddf.admin.configurator.OperationReport;
+import org.codice.ddf.admin.security.common.services.PolicyManagerServiceProperties;
 
 import com.google.common.collect.ImmutableList;
 
@@ -55,20 +54,22 @@ public class SaveWhitelistContexts extends BaseAction<ListField<ContextPath>> {
 
     @Override
     public ListField<ContextPath> performAction() {
+
         ListField<ContextPath> preUpdateWhitelistContexts = new ListFieldImpl<>(ContextPath.class);
-        for (String path :  getWhitelistContexts(configuratorFactory.getConfigReader())) {
+        for (String path : getWhitelistContexts(configuratorFactory.getConfigReader())) {
             preUpdateWhitelistContexts.add(new ContextPath(path));
         }
 
         Configurator configurator = configuratorFactory.getConfigurator();
-        configurator.updateConfigFile(POLICY_MANAGER_PID,
-                whiteListToPolicyManagerProps(contexts),
+        configurator.updateConfigFile(PolicyManagerServiceProperties.POLICY_MANAGER_PID,
+                new PolicyManagerServiceProperties().whiteListToPolicyManagerProps(contexts),
                 true);
 
         OperationReport configReport = configurator.commit(
                 "Whitelist Contexts saved with details: {}",
                 contexts.toString());
 
+        // TODO: tbatie - 4/23/17 - We need to discuss whether we want to return values even if it fails
         if (configReport.containsFailedResults()) {
             addMessage(failedPersistError());
             return preUpdateWhitelistContexts;
